@@ -1,12 +1,12 @@
 package hajiboot;
 
-import org.springframework.web.client.RestTemplate;
+import org.springframework.web.client.RestClient;
 
 public class Main {
 
 	public static void main(String[] args) {
-		RestTemplate restTemplate = new RestTemplate();
-		MarkdownRenderer markdownRenderer = new GitHubApiMarkdownRenderer(restTemplate);
+		RestClient.Builder restClientBuilder = RestClient.builder();
+		MarkdownRenderer markdownRenderer = new GitHubApiMarkdownRenderer(restClientBuilder);
 		MarkdownPrinter markdownPrinter = new MarkdownPrinter(markdownRenderer);
 		markdownPrinter.print(System.in);
 	}
